Refine contextual retry API implementation
