+++ fault-tolerance/fault-tolerance/src/main/java/io/helidon/faulttolerance/RetryConfigBlueprint.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, 2025 Oracle and/or its affiliates.
+ * Copyright (c) 2023, 2026 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,8 +26,9 @@ import io.helidon.builder.api.Prototype;
 /**
  * {@link Retry} configuration bean.
  */
-@Prototype.Blueprint(decorator = RetryConfigBlueprint.BuilderDecorator.class)
+@Prototype.Blueprint(decorator = FtBuilderSupport.RetryBuilderDecorator.class)
 @Prototype.Configured("fault-tolerance.retries")
+@Prototype.IncludeDefaultMethods({"jitterFactor", "maxDelay"})
 interface RetryConfigBlueprint extends Prototype.Factory<Retry> {
     /**
      * Default calls to make.
@@ -106,8 +107,10 @@ interface RetryConfigBlueprint extends Prototype.Factory<Retry> {
      * @return relative jitter factor
      */
     @Option.Configured
-    @Option.DefaultDouble(-1L)
-    double jitterFactor();
+    @Option.DefaultDouble(-1.0)
+    default double jitterFactor() {
+        return -1.0;
+    }
 
     /**
      * Maximum delay between invocation attempts, including jitter.
@@ -117,7 +120,9 @@ interface RetryConfigBlueprint extends Prototype.Factory<Retry> {
      * @return maximum delay, if configured
      */
     @Option.Configured
-    Optional<Duration> maxDelay();
+    default Optional<Duration> maxDelay() {
+        return Optional.empty();
+    }
 
     /**
      * Overall timeout of all retries combined.
@@ -165,78 +170,4 @@ interface RetryConfigBlueprint extends Prototype.Factory<Retry> {
     @Option.Configured
     @Option.DefaultBoolean(false)
     boolean enableMetrics();
-
-    class BuilderDecorator implements Prototype.BuilderDecorator<RetryConfig.BuilderBase<?, ?>> {
-        @Override
-        public void decorate(RetryConfig.BuilderBase<?, ?> target) {
-            validate(target);
-            if (target.name().isEmpty()) {
-                target.config()
-                        .ifPresent(cfg -> target.name(cfg.name()));
-            }
-            if (target.retryPolicy().isEmpty()) {
-                target.retryPolicy(retryPolicy(target));
-            }
-        }
-
-        /**
-         * Retry policy created from this configuration.
-         *
-         * @return retry policy to use
-         */
-        private Retry.RetryPolicy retryPolicy(RetryConfig.BuilderBase<?, ?> target) {
-            boolean jitterConfigured = !target.jitter().equals(Duration.ofSeconds(-1));
-            boolean jitterFactorConfigured = target.jitterFactor() != -1;
-            Retry.DelayingRetryPolicy.Builder delayBuilder = Retry.DelayingRetryPolicy.builder()
-                    .calls(target.calls())
-                    .delay(target.delay());
-
-            if (target.delayFactor() != -1) {
-                delayBuilder.delayFactor(target.delayFactor());
-            } else if (jitterConfigured || jitterFactorConfigured) {
-                delayBuilder.delayFactor(1);
-            }
-            if (jitterConfigured) {
-                delayBuilder.jitter(target.jitter());
-            }
-            if (jitterFactorConfigured) {
-                delayBuilder.jitterFactor(target.jitterFactor());
-            }
-            target.maxDelay().ifPresent(delayBuilder::maxDelay);
-            return delayBuilder.build();
-        }
-
-        private void validate(RetryConfig.BuilderBase<?, ?> target) {
-            if (target.calls() < 1) {
-                throw new IllegalArgumentException("Retry calls must be at least 1");
-            }
-            if (target.delay().isNegative()) {
-                throw new IllegalArgumentException("Retry delay must not be negative");
-            }
-            if (target.overallTimeout().isNegative() || target.overallTimeout().isZero()) {
-                throw new IllegalArgumentException("Retry overall timeout must be positive");
-            }
-            double delayFactor = target.delayFactor();
-            if (delayFactor != -1 && (!Double.isFinite(delayFactor) || delayFactor < 0)) {
-                throw new IllegalArgumentException("Retry delay factor must be -1 or a finite, non-negative number");
-            }
-            Duration jitter = target.jitter();
-            if (jitter.isNegative() && !jitter.equals(Duration.ofSeconds(-1))) {
-                throw new IllegalArgumentException("Retry jitter must be PT-1S or non-negative");
-            }
-            double jitterFactor = target.jitterFactor();
-            if (jitterFactor != -1 && (!Double.isFinite(jitterFactor) || jitterFactor < 0 || jitterFactor >= 1)) {
-                throw new IllegalArgumentException("Retry jitter factor must be -1 or from 0 (inclusive) to 1 "
-                                                           + "(exclusive)");
-            }
-            if (!jitter.equals(Duration.ofSeconds(-1)) && jitterFactor != -1) {
-                throw new IllegalArgumentException("Retry jitter and jitter factor cannot both be configured");
-            }
-            target.maxDelay().ifPresent(maxDelay -> {
-                if (maxDelay.isNegative()) {
-                    throw new IllegalArgumentException("Retry maximum delay must not be negative");
-                }
-            });
-        }
-    }
 }
