+++ fault-tolerance/fault-tolerance/src/main/java/io/helidon/faulttolerance/RetryContextImpl.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2026 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.faulttolerance;
+
+import java.time.Duration;
+import java.util.Objects;
+import java.util.Optional;
+
+record RetryContextImpl(int attempt,
+                        Duration elapsedTime,
+                        Duration previousDelay,
+                        Throwable previousThrowableValue) implements RetryContext {
+
+    @Override
+    public Optional<Throwable> previousThrowable() {
+        return Optional.ofNullable(previousThrowableValue);
+    }
+}
